+++ src/components/tags/tags.module.css
@@ -1,0 +1,58 @@
+.container {
+  min-width: 300px;
+  padding-left: 16px;
+  padding-right: 16px;
+}
+
+.sidebar {
+  display: flex;
+  flex-direction: column;
+  min-height: 56px;
+  max-width: 268px;
+  background-color: rgba(236, 73, 38, 1);
+  border-radius: 16px;
+  font-family: 'GolosText';
+  font-weight: 500;
+  font-size: 16px;
+  line-height: 24px;
+}
+
+.title {
+  width: 236px;
+  margin: 16px 16px 12px;
+}
+
+.tagList {
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-end;
+  flex-wrap: wrap;
+  padding: 0 16px 16px;
+  color: rgba(250, 250, 250, 1);
+}
+
+.button {
+  border: 1px solid rgba(250, 250, 250, 1);
+  padding: 4px 8px;
+  margin: 4px 0 0 4px;
+  background: rgba(236, 73, 38, 1);
+  font-size: 12px;
+  line-height: 16px;
+  border-radius: 100px;
+}
+
+.buttonActive {
+  border: 1px solid rgba(250, 250, 250, 1);
+  padding: 4px 8px;
+  margin: 4px 0 0 4px;
+  font-size: 12px;
+  line-height: 16px;
+  border-radius: 100px;
+  background: rgb(255, 255, 255);
+  color: rgb(0, 0, 0);
+}
+
+.button:active,
+button:focus {
+  outline: none;
+}
